Minor errors will suck your life away
I swear to god I was stuck with this thing hanging no matter what I did for like 6 hours. You know what the problem was?
if self.compare_item() == -1 or 0:
I CHECKED THIS BEFORE I WROTE IT
Literally went into the terminal to make sure I wasn't doing it wrong, and it worked. Except no, it didn't work, it only looked at the first part and completely ignored the other half without telling me. I've had so many issues in the past with this exact thing, because when I'd go in and tell it to compare multiple things for an if statement like if X == Y or X == Z or etc. it would throw it all out with an error.

But this time it's fine. Somehow this time it demands I write it like this:
if self.compare_item() == -1 or self.compare_item() == 0:
I'm upset.

I'm getting up early too to try and learn C++ with a friend.


I'm gonna hate life in the morning.